# openspec/specifications/apis/openapi.yaml
openapi: 3.0.3
info:
  title: AI Tool Usage Tracker API
  version: 1.0.0
  description: |
    REST API for the AI Tool Usage Tracker platform (Phase 1 MVP).

    **Authentication:** JWT Bearer token except `/auth/login`, `/auth/refresh`, and `/health`.

    **Errors:** RFC 7807 Problem Details (`application/problem+json`).

    **Traceability:** FR-PLT-001, ADR-012, NFR-SEC-003/004/006.

    ## RBAC Roles and Access

    | Role | Scope |
    |---|---|
    | `super_admin` | Full access to all endpoints and data across the entire organisation |
    | `team_admin` | Access to members, tools (read), teams (read), thresholds, uploads, collectors (read), and dashboard analytics — all scoped to managed teams only |
    | `finance_viewer` | Access to reports (read) and `GET /dashboard/my-usage` only |
    | `auditor` | Access to audit logs and `GET /dashboard/my-usage` only |
    | `team_member` | Access to `GET /dashboard/my-usage` only |
  contact:
    name: Platform Team
  license:
    name: Proprietary
servers:
  - url: https://api.ai-tool-tracker.example/api/v1
    description: Production
  - url: https://staging-api.ai-tool-tracker.example/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /roles:
    get:
      tags: [Roles]
      summary: List roles
      description: |
        Returns all roles for the organisation. Requires `settings` write permission (super_admin).
      operationId: listRoles
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/RoleListResponse'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
    post:
      tags: [Roles]
      summary: Create custom role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/CreateRoleRequest'
      responses:
        '201':
          description: Created role
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/RoleResponse'
        '409':
          $ref: './components/responses.yaml#/Conflict'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
  /roles/{role_id}:
    patch:
      tags: [Roles]
      summary: Update role
      operationId: updateRole
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/UpdateRoleRequest'
      responses:
        '200':
          description: Updated role
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/RoleResponse'
        '409':
          $ref: './components/responses.yaml#/Conflict'
        '404':
          $ref: './components/responses.yaml#/NotFound'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
    delete:
      tags: [Roles]
      summary: Delete custom role
      operationId: deleteRole
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role deleted
        '409':
          $ref: './components/responses.yaml#/Conflict'
        '404':
          $ref: './components/responses.yaml#/NotFound'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
  /roles/{role_id}/permissions:
    get:
      tags: [Roles]
      summary: Get role permission matrix
      operationId: getRolePermissions
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Permission matrix
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/PermissionMatrixResponse'
        '404':
          $ref: './components/responses.yaml#/NotFound'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
    put:
      tags: [Roles]
      summary: Replace role permission matrix
      operationId: replaceRolePermissions
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/ReplacePermissionsRequest'
      responses:
        '200':
          description: Updated matrix
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/PermissionMatrixResponse'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
        '404':
          $ref: './components/responses.yaml#/NotFound'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
...
  /reports/generate:
    post:
      tags: [Reports]
      summary: Generate report
      description: |
        Sync when `async: false` and result within 10s (NFR-PER-002).
        Returns 202 with `job_id` when async.

        **Roles:** Role-scoped read per report type.
      operationId: generateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/ReportGenerateRequest'
            examples:
              default:
                $ref: './components/examples.yaml#/ReportGenerateExample'
      responses:
        '200':
          description: Sync report — JSON body or file download
          content:
            application/json:
              schema:
                type: object
                description: Report data when format is json
            text/csv:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '202':
          description: Async report queued
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/ReportJob'
              examples:
                default:
                  $ref: './components/examples.yaml#/ReportJobExample'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
...
  /reports/jobs/{jobId}/download:
    get:
      tags: [Reports]
      summary: Download report artifact
      description: Redirects or returns presigned S3 URL (15 min expiry).
      operationId: downloadReportJob
      parameters:
        - $ref: './components/parameters.yaml#/ReportJobId'
      responses:
        '302':
          description: Redirect to presigned URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '404':
          $ref: './components/responses.yaml#/NotFound'
        '409':
          description: Job not completed
          content:
            application/problem+json:
              schema:
                $ref: './components/schemas.yaml#/ProblemDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/login. Required on all endpoints except /auth/login, /auth/refresh, /health.